Reorganize utility functions into tests/utils.c and testsuite/utils.c
